requirements.txt
flask
pandas
python-dotenv
openpyxl
supabase
gdown
requests
xlrd>=2.0.1
python-magic-bin
